(APWA-21) [AuthUserToRoot, AccountLinks, LinkList, NotAuthUserToLogin, SuggestedProducts, useBreadcrumbs, useImage, usePlaceholderImage, useResourceImage, useThumbnail, useProductQuantity, useProduct, useAtucomplete, useSearchBar, useSearchField, useSuggestions, useSearchPage, tests, style.]
- Refactored talons (SearchPage, SearchBar, Product, ProductQuantity, isSignedIn, Image, Breadcrumbs).
