+++ src/lib/components/LoadingIndicator/Loading-indicator.scss
@@ -1,3 +1,5 @@
+@import 'src/lib/styles/config/variables';
+
 .indicatorBlock {
     background-color: rgba(255, 255, 255, 0.5);
     bottom: 0;
