+++ src/lib/components/NotAuthUserToLogin/NotAuthUserToLogin.tsx
@@ -2,10 +2,10 @@ import React, { Fragment, FunctionComponent } from 'react';
 
 import RouterRoutes from 'src/lib/RouterRoutes/RouterRoutes';
 import { Redirect } from 'src/lib/drivers';
-import { useIsSignedIn } from 'src/peregrine/lib/talons/adeoweb/IsSignedIn/useIsSignedIn';
+import { useUserContext } from 'src/peregrine/lib/context/adeoweb/user';
 
 const NotAuthUserToLogin: FunctionComponent = ({ children }) => {
-    const { isSignedIn } = useIsSignedIn();
+    const [{ isSignedIn }] = useUserContext();
 
     if (!isSignedIn) {
         return <Redirect to={RouterRoutes.loginPage.url} />;
