--- src/lib/util/__mocks__/hooks/mockUseIsSignedIn.ts
@@ -1,5 +1,0 @@
-import { TUseSignIn } from 'src/peregrine/lib/talons/adeoweb/IsSignedIn/useIsSignedIn';
-
-const mock: TUseSignIn = { isSignedIn: false };
-
-export default mock;
