--- src/peregrine/lib/talons/adeoweb/Image/usePlaceholderImage.d.ts
@@ -1,13 +1,0 @@
-type TUsePlaceholderImageProps = {
-    displayPlaceholder: boolean;
-    imageHasError: boolean;
-    imageIsLoaded: boolean;
-};
-
-type TUsePlaceholderImage = {
-    shouldRenderPlaceholder: boolean;
-};
-
-export function usePlaceholderImage(
-    props: TUsePlaceholderImageProps
-): TUsePlaceholderImage;
