--- src/peregrine/lib/talons/adeoweb/IsSignedIn/useIsSignedIn.d.ts
@@ -1,5 +1,0 @@
-export type TUseSignIn = {
-    isSignedIn: boolean;
-};
-
-export function useIsSignedIn(): TUseSignIn;
