--- src/peregrine/lib/talons/adeoweb/IsSignedIn/useIsSignedIn.js
@@ -1,9 +1,0 @@
-import { useUserContext } from 'src/peregrine/lib/context/adeoweb/user';
-
-export const useIsSignedIn = () => {
-    const [{ isSignedIn }] = useUserContext();
-
-    return {
-        isSignedIn
-    };
-};
