--- src/peregrine/lib/talons/adeoweb/ProductImageCarousel/useThumbnail.d.ts
@@ -1,10 +1,0 @@
-type TUseThumbnail = {
-    handleClick: () => void;
-};
-
-interface IUseThumbnailProps {
-    itemIndex: number;
-    onClickHandler: (index: number) => void;
-}
-
-export function useThumbnail(props: IUseThumbnailProps): TUseThumbnail;
