--- src/peregrine/lib/talons/adeoweb/SearchBar/useSearchBar.d.ts
@@ -1,18 +1,0 @@
-import { MutableRefObject } from 'react';
-
-import { FormikProps } from 'src/lib/drivers';
-
-export type TUseSearchBarValues = {
-    search_query: string;
-};
-
-type TUseSearchBar = {
-    containerRef: MutableRefObject<null>;
-    initialValues: TUseSearchBarValues;
-    expanded: boolean;
-    setExpanded: (isExpanded: boolean) => void;
-    handleFocus: () => void;
-    form: FormikProps<TUseSearchBarValues>;
-};
-
-export function useSearchBar(): TUseSearchBar;
