--- src/peregrine/lib/talons/adeoweb/SearchBar/useSearchField.d.ts
@@ -1,5 +1,0 @@
-type TUseSearchFieldProps = {
-    setFieldValue: (field: string, value: string) => void;
-};
-
-export function useSearchField(props: TUseSearchFieldProps): void;
