--- src/peregrine/lib/talons/adeoweb/SearchBar/useSuggestions.d.ts
@@ -1,14 +1,0 @@
-import { TProductSearch } from 'src/lib/types/ProductSearch';
-
-type TUseSuggestionsProps = {
-    items: TProductSearch[];
-    setVisible: (isVisible: boolean) => void;
-    visible: boolean;
-};
-
-type TUseSuggestions = {
-    onNavigate: () => void;
-    shouldRender: boolean;
-};
-
-export function useSuggestions(props: TUseSuggestionsProps): TUseSuggestions;
